Update transport layer to use the right size_t size.

+++ sdk/core/azure-core/inc/azure/core/http/win_http_transport.hpp
@@ -32,8 +32,8 @@ namespace Azure { namespace Core { namespace Http {
 
   namespace _detail {
 
-    constexpr static int64_t DefaultUploadChunkSize = 1024 * 64;
-    constexpr static int64_t MaximumUploadChunkSize = 1024 * 1024;
+    constexpr static size_t DefaultUploadChunkSize = 1024 * 64;
+    constexpr static size_t MaximumUploadChunkSize = 1024 * 1024;
 
     struct HandleManager final
     {
